KS_2_1 исправлено

+++ src/main/kotlin/lessons2/Lesson2_Task1.kt
@@ -1,12 +1,13 @@
 package lessons2
 
 fun main() {
-    val student1 = 3
-    val student2 = 4
-    val student3 = 3
-    val student4 = 5
+    val student1 = 3f
+    val student2 = 4f
+    val student3 = 3f
+    val student4 = 5f
 
-    val countOfStudents: Float = 4f
+    val countOfStudents: Int = 4
+
 
     println("Средний балл:")
     println((student1 + student2 + student3 + student4) / countOfStudents)
